requirements.txt
pytest==6.2.5
pytest-bdd==3.2.1
snowflake-connector-python[pandas]==2.7.1
snowflake-sqlalchemy==1.3.2
SQLAlchemy==1.4.27